Parse the expires_at field from the installation token
It looks like the GitHub docs have changed here; it used to say that
the token was valid for 60 minutes, but now there's a field in the
response containing when it expires.

This commit just adds the code to parse that field (using chrono), but
doesn't actually use it yet.